Add functionality for mass deletion
Using hashmap_delete() with hashmap_scan() or hashmap_iter() is unsafe,
as a resize of the bucket list may happen at any time, throwing off the
current cursor.

However, mass deletion based on some part of the item that isn't the key
may still be useful in some cases.

This introduces voluntary tombstoning of buckets, with the possibility
to clean them up on demand.  Tombstoning a bucket is implemented as a
simple flag, which effectively makes the bucket still "occupied", but
otherwise inaccessible.  This can also be viewed as delayed deletion.

These two functions are added to make this possible:

- hashmap_unset (and hashmap_unset_with_hash), which simply sets the
  tombstone flag on the item.
- hashmap_vacuum, which goes through the whole bucket list, and deletes
  every bucket that has the tombstone flag.

Example code with hashmap_scan:

    bool filter_unset(const void *item, void *udata)
    {
        if (/* Some condition on |item| */)
            hashmap_unset(udata, item);
        return true;
    }
    // ...
    hashmap_scan(map, filter_unset, map);
    hashmap_vacuum(map);

Example code with hashmap_iter:

    for (size_t i = 0, void *item; hashmap_iter(map, &i, &item);) {
        if (/* Some condition on |item| */)
            hashmap_unset(udata, item);
    hashmap_vacuum(map);

A test is included.

Signed-off-by: Le <_@I-am-Le.me>